+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## [0.5.0] - 2021-06-09
 ### Added
 - Attribute Change from String
 
@@ -57,7 +58,8 @@ Minor fix in observer interface
 ## [0.1.0] - 2021-05-26
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.4.1...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.5.0
 [0.4.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.4.1
 [0.4.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.4.0
 [0.3.2]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.3.2